settlement v1.2 update

+++ src/main/java/com/bruceking/settlement/bean/Currency.java
@@ -1,10 +1,12 @@
 package com.bruceking.settlement.bean;
 
+import java.math.BigDecimal;
+
 public class Currency {
 
     private Integer currency_id;
     private String currency_type;
-    private Double currency_er;
+    private BigDecimal currency_er;
 
     public Integer getCurrency_id() {
         return currency_id;
@@ -22,11 +24,11 @@ public class Currency {
         this.currency_type = currency_type;
     }
 
-    public Double getCurrency_er() {
+    public BigDecimal getCurrency_er() {
         return currency_er;
     }
 
-    public void setCurrency_er(Double currency_er) {
+    public void setCurrency_er(BigDecimal currency_er) {
         this.currency_er = currency_er;
     }
 }
